included the utility to share some of the methods of server.py
